refactor: update namespaces and routes for MercadoLibre controllers
