+++ multistock-backend/routes/api.php
@@ -110,7 +110,7 @@ Route::post('/mercadolibre/save-credentials', [MercadoLibreController::class, 's
 // Generate MerccadoLibre login Auth 2.0 URL
 Route::post('/mercadolibre/login', [loginController::class, 'login']);
 // Handle MercadoLibre callback
-Route::get('/mercadolibre/callback', [MercadoLibreController::class, 'handleCallback']);
+Route::get('/mercadolibre/callback', [handleCallbackController::class, 'handleCallback']);
 // Check MercadoLibre connection status
 Route::get('/mercadolibre/test-connection/{client_id}', [testAndRefreshConnectionController::class, 'testAndRefreshConnection']);
 // Get MercadoLibre credentials if are saved in db
@@ -174,9 +174,6 @@ Route::get('/mercadolibre/compare-annual-sales-data/{client_id}', [compareAnnual
 
 // Get sales by date range
 Route::get('/mercadolibre/sales-by-date-range/{client_id}', [getSalesByDateRangeController::class, 'getSalesByDateRange']);
-
-// Get sales by date range
-Route::get('/mercadolibre/sales-by-date-range/{client_id}', [MercadoLibreDocumentsController::class, 'getSalesByDateRange']);
 
 // Info route
 Route::get('/info', [InfoController::class, 'getInfo']);
